requirements-client.txt
psutil
pyzmq==24.0.1
numpy
pillow

# to run the qt gui applications
pyside6==6.6.2

